requirements.txt
accelerate
transformers
diffusers[training]
datasets
peft
Pillow
torch